moved all unit tests to under src for proper folder discovery and such
