+++ src/tests/analysis/test_investigator.py
@@ -1,0 +1,1 @@
+import pytest
