+++ src/tests/test_entry_structures.py
