+++ src/tests/test_local_minima_variation.py
