+++ src/tests/test_weighted_structures.py
